Added a suid guacamayo-hostname utility for use by the system plugin
We need a way to allow the limited privilege user running Mex to change the
host name. We can't use the normal hostname utility because on busybox systems
this is just a link to the busybox binary, so we can't make it suid. So here
we go.

guacamayo-host name strips any control charasters from the mandatory argument
and then calls sethostname() and saves the new name in /etc/hostname to
persist across reboots.